# src/main/resources/swagger.yaml
swagger: '2.0'
info:
  title: Corporations
  version: 1.0.0
tags:
  - name : Corporation

paths:
  /corporation/inserir:
    post:
      tags:
        - corporation
      summary: Cria um novo Corporation.
      description: Cria um novo corporation com os dados fornecidos.
      operationId: insertCorporation
      parameters:
        - name: corporation
          in: body
          description: Dados do corporation a ser criado.
          required: true
          schema:
            $ref: '#/definitions/CorporationInput'
      responses:
        '201':
          description: Created
          schema:
            $ref: '#/definitions/Corporation'

  /corporation/getAll:
    get:
      tags:
        - Corporation
      summary: Obtém uma lista paginada de corporations.
      description: Retorna uma lista paginada de corporations.
      operationId: getAllCorporations
      parameters:
        - name: page
          in: query
          description: O número da página a ser retornada.
          required: false
          type: integer
          format: int32
          minimum: 0
          default: 0
        - name: size
          in: query
          description: O tamanho da página a ser retornada.
          required: false
          type: integer
          format: int32
          minimum: 1
          maximum: 100
          default: 20
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/PaginacaoDeCorporation'

definitions:
  TipoEnum:
    type: string
    description: 'teste'
    enum:
      - Financeiro
      - Agro
      - Tecnologia
  Corporation:
    type: object
    properties:
      id:
        type: integer
        format: int64
      nome:
        type: string
      tipo:
        type: string
        $ref: '#/definitions/TipoEnum'

  CorporationInput:
    type: object
    properties:
      nome:
        type: string
      tipo:
        type: string
        $ref: '#/definitions/TipoEnum'
        example: "Financeiro"
        description: "Tipo"
        pattern: "(()|[Financeiro]{10}|[Agro]{4}|[Tecnologia]{10})"
        x-pattern-message: "informe um valor igual a: Financeiro, Agro ou Tecnologia."
      situacao:
        type: "string"
        pattern: "(()|[0-1]{1})"
        x-pattern-message: "informe 0 ou 1."
      profissao:
        type: "string"
        example: "DESENVOLVEDOR"
        description: "Profissão do Cedente Pessoa Fisica"
        pattern: "(()|[AVALISTA]{8}|[DIRETOR]{7}|([DIRETOR]{7}.*[FINANCEIRO]{10})|([DIRETOR]{7}.*[GERAL]{5})|([DIRETOR]{7}.*[PRESIDENTE]{10})|([DIRETOR]{7}.*[SUPERINTENDENTE]{15})|([DIRETOR]{7}.*[VICE]{4}.*[PRESIDENTE]{10})|[PROCURADOR]{10}|[PROPRIETARIO]{12}|([PROPRIETARIO]{12}.*[DIRIGENTE]{9})|[SOCIO]{5}|([SOCIO]{5}.*[PROPRIETARIO]{12}))"
        x-pattern-message: "informe um valor igual a: AVALISTA, DIRETOR, DIRETOR FINANCEIRO, DIRETOR GERAL, DIRETOR PRESIDENTE, DIRETOR SUPERINTENDENTE, DIRETOR VICE PRESIDENTE, PROCURADOR, PROPRIETARIO,PROPRIETARIO DIRIGENTE, SOCIO, SOCIO PROPRIETARIO."
      nacionalidade:
        type: "string"
        example: "BRASILEIRO"
        description: "Nacionalidade"
        pattern: "(()|[BRASILEIRO]{10}|[OUTRO]{5})"
        x-pattern-message: "informe um valor igual a: BRASILEIRO ou OUTRO."
      estadoCivil:
        type: "string"
        example: "SOLTEIRO"
        description: "Estado Civil"
        pattern: "(()|[SOLTEIRO]{8}|[CASADO]{6}|[SEPARADO]{8}|[DIVORCIADO]{10}|[VIUVO]{5})"
        x-pattern-message: "informe um valor igual a: SOLTEIRO, CASADO, SEPARADO, DIVORCIADO ou VIUVO."
      uf:
        type: "string"
        example: "SP"
        description: "Sigla do Estado do endereco"
        minLength: 2
        maxLength: 2
        pattern: "([AC]|[AL]|[AP]|[AM]|[BA]|[CE]|[DF]|[ES]|[GO]|[MA]|[MT]|[MS]|[MG]|[PA]|[PB]|[PR]|[PE]|[PI]|[RJ]|[RN]|[RS]|[RO]|[RR]|[SC]|[SP]|[SE]|[TO]){2}"
        x-pattern-message: "informe um valor igual a: AC, AL, AP, AM, BA, CE, DF, ES, GO, MA, MT, MS, MG, PA, PB, PR, PE, PI, RJ, RN, RS, RO, RR, SC, SP, SE, TO."
      email:
        type: "string"
        example: "email@email.com"
        description: "Email"
        minLength: 5
        maxLength: 200
        pattern: "^(.+)@(.+)$"
        x-pattern-message: "informe um email válido no formato email@email.com."

  PaginacaoDeCorporation:
    type: object
    properties:
      corporations:
        type: array
        items:
          $ref: '#/definitions/Corporation'
      paginaAtual:
        type: integer
        format: int32
      totalDePaginas:
        type: integer
        format: int32